requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.9.0
